soap_service/requirements.txt
Flask==2.2.5
lxml==5.0.0
mysql-connector-python==8.1.0
urllib3==1.26.16
